Add a few missing tests
